requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.4
django-admin-thumbnails==0.2.6
django-chartjs==2.3.0
django-ckeditor==6.4.2
django-filter==22.1
django-jalali==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-phone-field==1.8.1
django-taggit==3.0.0
ghasedakpack==0.1.9
idna==3.3
jdatetime==4.1.0
Pillow==9.1.0
pycparser==2.20
pyfiglet==0.8.post1
pygame==2.1.2
PyMySQL==1.0.2
pytz==2021.1
requests==2.28.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.1
suds-py3==1.4.5.0
tzdata==2022.1
urllib3==1.26.9
django-jazzmin==2.6.0
djangorestframework==3.14.0

